Move progress icon
