+++ Assets/Scripts/Lootables/CanisterHUD.cs
@@ -2,10 +2,11 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using DeCrawl.UI;
 
 namespace YLHalf
 {
-    public class CanisterHUD : HUDProgressionIcon
+    public class CanisterHUD : ProgressionIcon
     {
         public Canister canister;
