+++ Assets/deCrawl/Scrips/World/ILevel.cs
@@ -6,6 +6,7 @@ namespace DeCrawl.World
 {    public interface ILevel<Entity, ClaimCondition>
     {
         public bool ClaimPosition(Entity claimer, Vector3Int position, ClaimCondition condition);
+        public bool ClaimPositionForced(Entity claimer, Vector3Int position);
 
         public Vector3Int AsWorldPosition(Vector3Int gridPosition);
         public Vector3Int AsGridPosition(Vector3 worldPosition);
